requirements.txt
watchdog
GitPython